Añadidas prubeas unitarias y manejo de excepciones

+++ src/main/java/mx/uam/ayd/proyecto/negocio/ServicioDocumento.java
@@ -9,7 +9,6 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import lombok.extern.slf4j.Slf4j;
 import mx.uam.ayd.proyecto.datos.RepositoryDocumento;
 import mx.uam.ayd.proyecto.negocio.modelo.Documento;
 import mx.uam.ayd.proyecto.negocio.modelo.SolicitudTramite;
@@ -19,7 +18,6 @@ import mx.uam.ayd.proyecto.negocio.modelo.SolicitudTramite;
  * 
  * @author Adolfo Mejía
  */
-@Slf4j
 @Service
 public class ServicioDocumento {
 
@@ -28,36 +26,65 @@ public class ServicioDocumento {
 
     /**
      * Comunica al repositorio que debe eliminar un documento
+     * 
      * @param documento el documento a eliminar
+     * @throws IllegalArgumentException
      */
-    public void delete(Documento documento) {
+    public void delete(Documento documento) throws IllegalArgumentException {
+        if (documento == null) {
+            throw new IllegalArgumentException("Parametro nulo inválido");
+        }
         documentoRepository.delete(documento);
     }
 
-    
     /**
-     * Elimina los documentos adjuntados por el agremiado de una solicitud de trámite y establece en
+     * Elimina los documentos adjuntados por el agremiado de una solicitud de
+     * trámite y establece en
      * su lugar una lista vacía
-     * @param solicitudSeleccionada la solicitud de la que se quieren eliminar dichos documentos
+     * 
+     * @param solicitudSeleccionada la solicitud de la que se quieren eliminar
+     *                              dichos documentos
+     * @throws IllegalArgumentException
      * @return la solicitud actualizada
      */
-    public SolicitudTramite eliminarDocumentos(SolicitudTramite solicitudSeleccionada) {
-        List<Documento> documentosParaRechazar = solicitudSeleccionada.getRequisitos();
-        for (Documento documento : documentosParaRechazar) {
-            delete(documento);
+    public SolicitudTramite eliminarDocumentos(SolicitudTramite solicitudSeleccionada_)
+            throws IllegalArgumentException {
+        SolicitudTramite solicitudSeleccionada = solicitudSeleccionada_;
+
+        if ((solicitudSeleccionada == null) || (solicitudSeleccionada.getRequisitos() == null))
+            throw new IllegalArgumentException("Solicitud nula o sin documentos");
+
+        try {
+
+            List<Documento> documentosParaRechazar = solicitudSeleccionada.getRequisitos();
+            for (Documento documento : documentosParaRechazar) {
+                delete(documento);
+            }
+            solicitudSeleccionada.setRequisitos(new ArrayList<Documento>());
+            return solicitudSeleccionada;
+
+        } catch (IllegalArgumentException e) {
+            throw e;
         }
-        solicitudSeleccionada.setRequisitos(new ArrayList<Documento>());
-        return solicitudSeleccionada;
     }
 
-    
     /**
-     * Crea un objeto de tipo Documento en base a la dirección de un archivo pdf seleccionada por el usuario
+     * Crea un objeto de tipo Documento en base a la dirección de un archivo pdf
+     * seleccionada por el usuario
+     * 
      * @param pathDocTramiteFinalizado la dirección del archivo pdf
-     * @param tipoDocumento el tipo de documento que se está creando
+     * @param tipoDocumento            el tipo de documento que se está creando
+     * @throws IOException
+     * @throws IllegalArgumentException
      * @return el objeto de tipo Documento creado
      */
-    public Documento creaDocumento(Path pathDocTramiteFinalizado, String tipoDocumento) {
+    public Documento creaDocumento(Path pathDocTramiteFinalizado_, String tipoDocumento_) throws IOException {
+
+        Path pathDocTramiteFinalizado = pathDocTramiteFinalizado_;
+        String tipoDocumento = tipoDocumento_;
+
+        if ((pathDocTramiteFinalizado == null) || (tipoDocumento == null))
+            throw new IllegalArgumentException("Argumento nulo no válido");
 
         Documento documentoTramiteFinalizado = new Documento();
         documentoTramiteFinalizado.setTipoDocumento(tipoDocumento);
@@ -66,13 +93,12 @@ public class ServicioDocumento {
 
             byte[] bytesDocTramiteFinalizado = Files.readAllBytes(pathDocTramiteFinalizado);
             documentoTramiteFinalizado.setArchivo(bytesDocTramiteFinalizado);
+            return documentoTramiteFinalizado;
 
         } catch (IOException e) {
-            log.info("Error: " + e.getMessage());
-            return null;
+            throw e;
         }
 
-        return documentoTramiteFinalizado;
     }
 
 }
